Split adapters.h into per-version header files
